improve decryption
EncryptedSongReader will now reuses a cipher.BlockMode, and is now
twice as fast at decrypting. ReadChunk will also now return
the proper amount of bytes read with ReadFull. This fixes the
"corruption" at the end of decrypted songs, because with the old
behavior, there would be null bytes added to the end of the decrypted
song until the size of it was a multiple of 2048. Now decrypted and
encrypted files are the same size, as they should be.